ServerSideSavingPlugin: added host to removeTiddler's context
In TiddlySpace, a DELETE triggers a subsequent GET (against the recipe),
leading to the resulting tiddler missing the server.host field.

While it's possible that this could be addressed further down the stack
(by using the respective tiddler object's server.host field as fallback
if context.host is not available), this seems more in line with how the
adaptor mechanism is supposed to work in TiddlyWiki.

git-svn-id: http://svn.tiddlywiki.org/Trunk@12450 bb0f57cd-c710-0410-a8fb-e8bc1be0d679