Revert "TiddlyWebAdaptor: added support for diverging tiddler titles"
Using server.title as a valid persistent identifier to a resource on the
server leads to ambiguity (and thus subtle bugs), as this field is
already being used to detect renames.

This ambiguity is particularly harmful in moveTiddler, which triggers
getTiddler (to retrieve the ETag) - yet each method used a different
meaning for server.title.

In the future, a different field (server.id?) might be used instead to
support tiddler titles which differ between client and server.

This reverts commit 66c3dd4696d83f048a9b953a6820cdb75e60c9f2
(changeset:12453).

git-svn-id: http://svn.tiddlywiki.org/Trunk@12460 bb0f57cd-c710-0410-a8fb-e8bc1be0d679